+++ app/authz/authentication.go
@@ -1,9 +1,11 @@
 package authz
 
 import (
+	"app/authz/token"
 	"app/authz/token/storage"
 	"app/db"
 	"app/models"
+	"encoding/json"
 	"errors"
 
 	"github.com/gomodule/redigo/redis"
@@ -36,12 +38,55 @@ func NewAuthLayer(data *db.DataStorage, redis redis.Conn) AuthLayerInterface {
 	}
 }
 
-func (al *AuthLayer) AuthenticateUser(email string, password string) (string, error) {
-	return "", nil
+func NewAuthLayerMock(data *db.DataStorage) AuthLayerInterface {
+	tokenStorage := storage.NewInMemoryTokenStorage()
+	return &AuthLayer{
+		tokenStorage,
+		data,
+	}
 }
 
-func (al *AuthLayer) GetTokenData(token string) (*TokenData, error) {
-	return nil, nil
+func (a *AuthLayer) AuthenticateUser(email string, password string) (string, error) {
+	user, err := a.dataStorage.Users.GetUserByEmail(email)
+	if err != nil {
+		return "", err
+	}
+
+	passwordCheckError := checkUserPassword(password, user.HashedPassword)
+	if passwordCheckError != nil {
+		return "", ErrInvalidPassword
+	}
+
+	bytes, err := json.Marshal(&TokenData{
+		UserData: user,
+	})
+	if err != nil {
+		return "", ErrInvalidToken
+	}
+
+	token := token.NewToken()
+	if err = a.tokenStorage.SetTokenData(token, string(bytes)); err != nil {
+		return "", err
+	}
+
+	return token, nil
+}
+
+func (a *AuthLayer) GetTokenData(token string) (*TokenData, error) {
+	if len(token) == 0 {
+		return nil, ErrInvalidToken
+	}
+
+	tokenDataString, err := a.tokenStorage.GetTokenData(token)
+	if err == storage.ErrInvalidToken {
+		return nil, ErrInvalidToken
+	}
+
+	var tokenData TokenData
+	if err = json.Unmarshal([]byte(tokenDataString), &tokenData); err != nil {
+		return nil, err
+	}
+	return &tokenData, nil
 }
 
 func (al *AuthLayer) TokenStorage() storage.TokenStorage {
